S08P22A101-287 [:data:book]

Closes S08P22A101-287

See merge request s08-ai-image-sub2/S08P22A101!25

+++ ItsGrey/domain/repo/src/main/java/com/tntt/repo/BookRepository.kt
@@ -5,7 +5,7 @@ import com.tntt.home.model.SortType
 import com.tntt.model.BookInfo
 
 interface BookRepository {
-    fun getBookInfos(userId: String, sortType: SortType, startIndex: Int, bookType: BookType): List<BookInfo>
+    fun getBookInfos(userId: String, sortType: SortType, startIndex: Long, bookType: BookType): List<BookInfo>
     fun createBook(userId: String): String
     fun deleteBook(bookIds: List<String>): Boolean
